Break AstClickhouseQuery into its own file
